WIP: Support for PixelFormat::BGR565_sRGB. Probably fixed all but extensions (gfxdevices etc).

+++ src/wondergui/base/wg_msgrouter.cpp
@@ -244,7 +244,7 @@ namespace wg
 
 	int MsgRouter::clearMsgQueue()
 	{
-		int nMessages = m_msgQueue.size();
+		int nMessages = (int) m_msgQueue.size();
 		m_msgQueue.clear();
 		return nMessages;
 	}
